extension: Full rework
We should use the ConnMan D-Bus API more closely to its intent,
instead of merely trying to bolt it on top of our class herarchy.

The structure has been changed to:

  CMApplet
    CMTechnologyTitleMenuItem('ethernet')
      CMServiceMenuItem → CMService
      ...
    CMTechnologyTitleMenuItem('wifi')
      CMServiceMenuItem → CMService
      ...

The CMApplet is our main container; its icon reflects the state
property of ConnMan, while its internal logic controls the menu
hierarchy for each technology available and enabled. State and
property changes update the icon and rebuild the menu hierarchy.

The CMTechnologyTitleMenuItem controls the main technology section,
with the switch for enabling/disabling the technology; it also has
the control for each service used by the technology type.

The CMServiceMenuItem displays a CMService, which matches a ConnMan
service object.

This separation matches more closely the object system provided by
ConnMan itself.