Use the service state, not the global one
The net.connman.Manager:State property is pretty much useless if
we want to update the icon in the panel, as it doesn't give us
the multiplexed state, but just a binary 'online/offline' state.
We can use the CMService::changed signal to update the icon
instead, by iterating through the services list.